requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
numpy>=1.24
opencv-python-headless>=4.8
pillow==12.1.1
pytesseract>=0.3.10
python-dotenv==1.0.1
pytest==8.3.5
pytest-flask==1.3.0
